Develop

+++ workspace/apps/api/src/app/board/board.service.ts
@@ -1,10 +1,13 @@
-import { Injectable, NotFoundException, BadRequestException } from '@nestjs/common';
-import { CreateCommentDto, CreatePostDto, UpdatePostDto } from '../common/dtos/post.dto';
+import { Injectable, NotFoundException, BadRequestException, UnauthorizedException } from '@nestjs/common';
+import { CreateCommentDto, CreatePostDto, UpdatePostDto, UpdateCommentDto } from '../common/dtos/post.dto';
 import { PrismaService } from '../prisma/prisma.service';
+import { UserService } from '../user/user.service';
 
 @Injectable()
 export class BoardService {
-    constructor(private prisma: PrismaService) {}
+    constructor(
+        private prisma: PrismaService,
+        private userService: UserService) {}
 
     async getPostAll() {
         try {
@@ -23,7 +26,9 @@ export class BoardService {
             const response = await this.prisma.post.findUnique({
                 where: {
                     id: postId
-                }
+                },
+
+                include: { comments: true }
             });
 
             return response;
@@ -48,37 +53,49 @@ export class BoardService {
         }
     }
 
-    async updatePost(postId: string, updatePostDto: UpdatePostDto) {
+    async updatePost(postId: string, updatePostDto: UpdatePostDto, userId: string) {
         try {
-            const id = postId;
-            const updateData = {
-                slug: updatePostDto.slug,
-                title: updatePostDto.title,
-                body: updatePostDto.body,
-                published: updatePostDto.published
-            };
+            const postObject = await this.getPostById(postId);
+
+            if (!postObject) {
+                throw new BadRequestException('INVALID postId');
+            }
+
+            if (postObject.userId !== userId) {
+                throw new UnauthorizedException('INVALID userId');
+            }
 
             const response = await this.prisma.post.update({
                 where: {
-                    id: id
+                    id: postId
                 },
-                data: updateData
+                data: {
+                    ...updatePostDto
+                }
             });
 
             return response;
         }
-        catch(e) {
-            throw new BadRequestException(e);
+        catch(err) {
+            throw new BadRequestException(err);
         }
     }
 
-    async deletePost(postId: string) {
+    async deletePost(postId: string, userId: string) {
         try {
-            const id = postId;
+            const postObject = await this.getPostById(postId);
+
+            if (!postObject) {
+                throw new BadRequestException('INVALID postId');
+            }
+
+            if (postObject.userId !== userId) {
+                throw new UnauthorizedException('INVALID userId');
+            }
 
             const response = await this.prisma.post.delete({
                 where: {
-                    id: id
+                    id: postId
                 }
             });
 
@@ -105,13 +122,65 @@ export class BoardService {
         }
     }
 
+    async getComment(postId: string) {
+        try {
+            const response = await this.prisma.comment.findMany({
+                where: {
+                    postId: postId
+                }
+            });
+
+            return response;
+        }
+        catch(e) {
+            throw new BadRequestException(e);
+        }
+    }
+
     async createComment(createCommentDto: CreateCommentDto, userId: string) {
         try {
-            const userName = '';
+            const profileObject = await this.userService.getUserProfileById(userId);
+
+            const userName = profileObject.name;
+
             const response = await this.prisma.comment.create({
                 data: {
                     ...createCommentDto,
-                    name: userName
+                    userName: userName,
+                    userId: userId
+                }
+            });
+
+            return response;
+        }
+        catch(e) {
+            throw new BadRequestException(e);
+        }
+    }
+
+    async updateComment(commentId: string, updateCommentDto: UpdateCommentDto, userId: string) {
+        try {
+            const response = await this.prisma.comment.update({
+                where: {
+                    id: commentId
+                },
+                data: {
+                    ...updateCommentDto
+                }
+            });
+
+            return response;
+        }
+        catch(e) {
+            throw new BadRequestException(e);
+        }
+    }
+
+    async deleteComment(commentId: string) {
+        try {
+            const response = await this.prisma.comment.delete({
+                where: {
+                    id: commentId
                 }
             });
